fix: preserve style set export extensions
